+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
+### Added
+- Equality operator for `StitchCount` class
+
 ## [0.2.0] - 2026-03-09
 
 ### Added